[media] i2c: uh2c: Read the CSI bus config from DT
The driver is basically ready to support less than 4 lanes, we only
need to extend the g_mbus_config callback to properly report the
number of lanes set in the DT.

While at it also read the clock flags from DT even if those are not
yet used by the driver. To stay compatible with the old code the DT
should define a clock lane, but not set the clock-noncontinuous flag.

Change-Id: I8ab9fc5854aaa5e21e9852fb6b169a2021fc1b9d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5602
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>