[media] v4l2-async: Always unregister the subdev on failure
In v4l2_async_test_notify() if the complete notifier returns an error
the subdev is not unregistered. This leave a path where
v4l2_async_register_subdev() can fail but leave the subdev still
registered.

Add the required call to v4l2_device_unregister_subdev() to plug
this holes.

Change-Id: Iabbfdf78d6f0a3bce0497b65b358a7ba24028e5c
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5609
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>