[media] i2c: uh2c: Add a control to read the audio sample rate
Alsa expects the codec to be able to select any of the supported sample
rates and has no provision to query the current rate. As the chip can't
resample the audio signal the user application must be able to query the
current sample rate, this is now possible with this new control.

Change-Id: I2c07008f5e2080b61e3b28dab2ae18f793bdfbd9
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5614
Reviewed-by: Julian Scheel <julian@jusst.de>