video: tegra: hdmi: Fix the initial audio setup
After the recent changes to the HDMI state machine to avoid a display
reset during boot, it turns out that the audio is then not working.
This is because the early HDMI init doesn't take care of the audio
setup. To fix this extend the EDID recheck state of the HDMI state
machine to setup the audio if it hasn't been done yet.

Change-Id: Idfc77bb8571f3921fe8f432324ad4b19ac5b5c2f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5615
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>