ARM: tegra: meerkat: Configure and enable cl-dvfs
The cl-dvfs configuration is copied from Laguna platform data, which is
what Jetson uses, with one exception: for laguna, the vdd map, which
maps target voltages to register values that are poked into the pmu
register driving the cpu voltage, is filled in by board code.

For meerkat this is left to the cl-dvfs driver, which queries the cpu
regulator for the output voltages it supports. Specifying a conversion
slope of 1, and, implicitly, a zero conversion offset simply generates
a list of register values starting from 0 with an increment of 1.

Change-Id: I7c8a077fbe1b76a21f81e7937d493778c886af53
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5912
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>