ARM: tegra: meerkat: Add dvfs board code and platform data
In particular, this enables soc-therm and balanced throttling on
meerkat.  The platform data is copied from Jetson.

We would prefer to have at least the thermal trip points and their
connection to cooling devices being configured in the device-tree rather
than by board code.  But unfortunately, the dvfs configuration data
structures are quite deeply integrated into the tegra platform drivers,
and their mapping of cooling devices and thermal trip points does not
match that of the standard thermal framework.

So, until we are running into the strict requirement to handle multiple
dvfs configurations for meerkat platforms, let's start with a common one
that's implemented with board code.

Change-Id: Ic7bcc0a7d116ba65615b30042ffc134d6368b75b
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5913
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>