[media] imx219: Code cleanup
In the s_ctrl handler, move the range checks on the user-provided value
to the handlers of the individual controls. Since this makes a comment
about the maximum gain at the handler redundant, it's removed. Also,
simplify the error paths of the control handlers and the chip ID
checking code a bit.

Change-Id: Iedb9de1e6042db85dbc1997789809ef6ce978e8c
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5817
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>