[media] tegra: vi2: Fix the frame sizes and intervals enumeration
The frame sizes and intervals enumeration ioctls pass a pixel format
to the video device, which is just passed down to the sensor
subdevice, however subdevices works with mbus formats and not pixel
formats. Fix both ioctl to convert the pixel format to an mbus format
before passing it down to the subdevice.

Change-Id: Iccba29b936bd7028f9732201395b8cc712d636b7
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5814
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>