video: tegra: hdmi: Implement proper early framebuffer init
Allow an early and flicker-free framebuffer init for hdmi displays that
preserves state when transitioning to the regular hdmi state machine.

The existing code that tried to achieve this was not initializing the
display state completely.  It did not initialize the framebuffer mode
list from the early edid read, required extra code to init audio in the
hdmi state machine, did not implement the complete logic to find the
preferred video mode, and was missing other pieces.

To remedy this, add a new hook .get_monspecs to tegra_dc_out_ops that is
called early when the dc is probed.  This can be used to query a
connected display early for supported video modes and identify a
preferred mode before the dc is enabled and a framebuffer is available.
This allows to set the preferred mode before enabling the dc, avoiding
screen blanking when the video mode is reset.  Also the framebuffer can
be initialized with a valid modelist.  To bridge the gap between the
call to .get_monspecs and the framebuffer init, the monspecs and
modelist are temporarily saved to the framebuffer platform data, and
merged into the framebuffer state when the framebuffer is registered.

Add an implementation of .get_monspecs for hdmi that queries the display
edid and initializes the display state.

As an extra benefit, introducing the new hook kills some hdmi specific
code from the dc driver, moving the necessary bits to the hdmi driver.
It also kills the extra audio init code in the hdmi state machine, which
is now no longer required.

Change-Id: I554b17376a5b21f93afa94902059b1bba08b30b5
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5964
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>