video: tegra: hdmi: Fix test for early edid data
When the hdmi state machine is starting, edid data may already have been
retrieved even if the dc was not enabled early.  Modify the check for
existing edid data to test the flag that signals complete processing of
the eld data.  Temporarily grabbing the raw edid data itself is not
necessary.

Change-Id: Ic947962e73ea0010028e22a4ae8c4b1cfc983ac5
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5965
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>