video: tegra: hdmi: Foster the monspec apply code a bit
Replace pr_info and friends with dev_*. Since the eld_retrieved flag is
used to signal successful edid processing, set it only after the entire
processing has successfully completed.  Now that we have real early hdmi
init in the dc, reword a comment slightly so that it is still accurate.

Change-Id: Ia5267b13e20f3ff94a7a4b6c1a23800c05d12d46
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5969
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>