video: tegra: lvds: Fix the clock configuration
LVDS uses the SOR for its output, however unlike the DC the SOR doesn't
have any divider. As such it must be driven by a clock running at the
same rate as the pixel clock. Rework the clock setup to have the DC
and SOR driven by the same clock running at the pixel clock rate.

Change-Id: I36eb034cebe2e64b5ef489813f5aa3ba5e592b15
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6034
Reviewed-by: Julian Scheel <julian@jusst.de>