pwm-backlight: Revert all the nvidia hacks
nvidia added various dubious hacks to this driver and reverted most of
them later on. This patch reverts what was left to bring this file back
to the version from the last mainline commit.

Commit 95730406a0e3bb (video: backlight: pwm_bl: add pwm_gpio) added a
new field to the platform data, this is not reverted by this patch to
avoid having to fix the platform data.

Change-Id: I884469e0e4f6fda6bdb3f68a926088eb40da81cb
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6012
Reviewed-by: Julian Scheel <julian@jusst.de>