ARM: tegra12: clock: Remove the SW limits on PLLs
PLL RE, SS, M and XC all had a trivial implementation of the
round_p_to_pdiv() callback that only coped with the first few possible
values of p. It seems p had been limited to the range where there is a
linear mapping between p and pdiv.

Fix this by adding a common implementation similar to the one used for
PLL CX: just go thru the array and return the first value greater or
equal to the requested one.

Change-Id: I24fd92bee33da4e0f2de0caec54b01e9894fada9
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6116
Reviewed-by: Julian Scheel <julian@jusst.de>