ARM: tegra: meerkat: Configure pll_c to allow using it for LVDS
pll_c is normally used for sdmmc4 (the internal eMMC), but it is the
only usable pll when HDMI and LVDS are needed at the same time.
Currently pll_c run at 600MHz and the sdmmc4 use a 200MHz clock,
however when pll_c is running at such a rate the DVFS doesn't properly
cope with the switch to the pixel clock rate needed for LVDS.

As a work-around we configure pll_c for 200MHz, that way the sdmmc4
still run at 200MHz when no LVDS is used and the DVFS doesn't explode
when LVDS is used.

Change-Id: If6d2dde079a1c008858573980b3cbf30ee9d62f9
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6117
Reviewed-by: Julian Scheel <julian@jusst.de>