ARM: tegra: cpuquiet: Delay the initial start
When cpuquiet starts too early it might happen that the system is
running on the LP core when the core are counted for the boot logo. In
this case only one penguin appear instead of 4. To prevent this, as
well as making sure that the boots process is as fast as posible, delay
the cpuquiet starts by 10 seconds.

Change-Id: Ia6d4d2e01650a72d587ca48a36d4097a80bda423
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6118
Reviewed-by: Julian Scheel <julian@jusst.de>