asoc: tegra: wm8903: Fix probe() error handling and remove order
The probe() error handling was not correct as the un-init was not in
reverse sequence of the init. To fix this and simplify the code move
the headset switch registration at the end of the probe, it allow to
replace most goto by simple return. Finally also adjust the remove()
function to unregister in the correct order.

Change-Id: I1b076c2b4ee7d07a999c801e23a4da9172c7dc8d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>