usb: tegra: Fix the EHCI platform data parsing from DT
Both the EHCI and OtG driver need access to the EHCI platform data. To
solve this we use a generic parsing function, tegra_usb_parse_dt(),
which can be called from both driver. However this function use devres
on the EHCI device, even if the EHCI device has not yet been probed.
This lead to a warning when the EHCI device is probed, but it is also
a problem if the EHCI device is removed. In the latter case devres
would free the pdata but the OtG device would still hold a pointer to
it.

To solve this just kzalloc() the EHCI pdata, later calls to
tegra_usb_parse_dt() will do nothing and the pdata will always stays
available.

Change-Id: Id876914771e9bfd99f8ffae6698ad9f490457c83
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6063
Reviewed-by: Julian Scheel <julian@jusst.de>