iio: as3722-adc: Convert driver from hwmon to iio
So far the driver has exposed only a subset of the adc channels, which
are covered by the hwmon use-case.  Convert the driver from hwmon to
iio, and add support for the general-purpose channels of the adc.

For the gpio and pwm pins, only low-voltage up to 1.56mV is supported so
far. Support for reading the current for sd0, sd1 and sd6 is not there
yet since it depends on an external shunt resistor. The driver can be
easily extended to overcome these limitations though.

Change-Id: Id0d9c4b24669e89cb300bb3aa415c03b06201d24
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6141
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>