pinctrl: as3722: Don't re-use the AS3722_GPIO_MODE prefix
The AS3722_GPIO_MODE prefix was used for 2 different things. First for
the values that can be set in the gpioX_mode register and then for
another set of defines that code the pinctrl properties set on a GPIO.
This is not only confusing it also led to a bug because a value
intended for the register was set in the property field.

To fix this mess we rename the properties values to use the new
AS3722_GPIO_CONFIG prefix, reflecting the PIN_CONFIG prefix of the
pinctrl API. To still get readable code we also rename the field
where this value is stored from mode_prop to config_prop. Finally we
also make sure to always use 'unsigned' as type when handling the
properties values.

Change-Id: I55ea7cc29350b4aff6e48619eebffb2c484b4297
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6122
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>