pinctrl: as3722: Read the initial pin config from the hardware
Currently the probe only reads the GPIO invert setting from the chip,
extend it to also read the pull settings and function. This makes sure
that we have a consistent state between hardware and software.

We also change the io_function field from storing the pinmux function
to storing the equivalent register value. This member is used to check
that the configured function is correct when requesting a GPIO. For
the GPIO function both the pinmux function and the register value are
0, so this single use doesn't need to be changed. This change is mostly
to save us from adding a conversion from register value to pinmux
function.

Change-Id: If6d75477ed0badb4c7aacb3d0eec73774862f367
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6130
Reviewed-by: Julian Scheel <julian@jusst.de>