hwmon: iio: add label for channels read by iio_hwmon
This adds a label property for each channel, the label can be defined
in DT with the linux,hwmon-label property or is taken from the
datasheet name of the IIO channel. The DT property can be used with
generic IIO devices to give a label that describes the specific use
of the IIO channel on the board.

Change-Id: Ic703e5910a98a9367aecfc6c19a0c24fccace019
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6164
Reviewed-by: Julian Scheel <julian@jusst.de>