[media] tegra: vi2: Fix get/enum format whithout sensors
Currently these functions return success with the data set to all zero
when no sensor is available. Fix this by using the
tegra_vi_channel_input_lock/unlock() functions that also check for the
sensor. While at it fix these functions to return -EINVAL when the
requested buffer type is not correct.

Change-Id: I689a345f2f6b1edf446c3ca08af0ec77c89dc8dc
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6388
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>