ARM: tegra: meerkat: Fix the video output clocking
Currently we use pll_d2 for the HDMI output and pll_c for the
secondary displays. To allow LVDS to run from pll_c we lowered it from
600MHz to 200MHz as it looked like that was sufficent to clock the
SDMMC and a secondary display. However it turns out that pll_c is also
used for the EMC to run at 600MHz and 300Mhz and as backup clock for a
few other high frequency. This killed the performances as the EMC
wasn't able to run above 528Mhz.

However it turns out that Jetson use pll_d_out0 for the primary
display. This clock is also used for CSI and DSI so it can't be used
for LVDS as LVDS need a PLL set to the pixel clock rate of the display.
But it is fine for HDMI and that leave pll_d2 free for the secondary
display.

To fix this mess we now use pll_d_out0 as default PLL for HDMI on the
Meerkat SOM and switch the secondary displays to pll_d2. This allow
us to leave pll_c at 600MHz and still support dual display
configurations. As an exception we still use pll_d2 for HDMI on the
Kitt and Eve boards as these use HDMI as a secondary display.

Change-Id: I687a02b656063655cc9508cbed9cd0f223798eb1
Fixes: 5a11ae99e4b1 ("ARM: tegra: meerkat: Configure pll_c to allow using it for LVDS")
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6404
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>