gitattributes: Add template file for storing commit info into archives
Drop a new template file `git-export.commit' into the top-level
directory with commit information placeholders that get expanded via an
export-subst gitattribute when the tree is archived.

This allows precise version identification of an archived kernel tree -
which does not hold any git history any more - if the existing, standard
version numbers are insufficient.

For this, the top-level .gitignore need to be amended to allow tracking
of .gitattributes files.

Change-Id: Id2a089b98ec0cb533d0f6c6c79cfe8507376256b
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6260
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>