ARM: tegra: meerkat: Don't specify vdd_3v3_run gpio
Remove the vdd_3v3_run gpio description to avoid a reference loop
between as3722 regulator, core regulator and as3722 pinmux drivers,
which causes a probe deferral on early regulator init. This is to be
avoided to allow the kernel to disable unused regulators on bootup.
Removing the gpio specification in here effectively is a no-op as the
supply is enabled before the kernel starts and marked as always-on, so
that it cannot be disabled at runtime.

Change-Id: I640276a8432548caf6094489d64bf98cd4776744
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/6864
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>