ARM: tegra: meerkat: Move the baseboard regulators before the PMIC
The baseboards have to provide a 3.3V and 5V supply for the Meerkat COM.
However as they then appear after the PMIC regulators in the DTB this
leads to a lot of probe deferrals. This in turn prevents the kernel from
disabling unused supplies on the PMIC.

To solve this we move the COM regulators to a separate node and add
an empty 'regulators' nodes for the baseboards. In this way the
baseboard's regulators appear before the COM regulators in the DTB, but
we can still define them in the baseboard DTS.

Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Change-Id: I01aa08ead7d4f07b6a979e50414a2d2ea4204bfa
Reviewed-on: http://review.adnet.avionic-design.de/6863
Reviewed-by: Julian Scheel <julian@jusst.de>