ata: ahci-tegra: Fix various bugs in the regulators handling
The tegra_ahci_controller_init() function, among other things, take
care of getting and enabling the regulators. However it would always
disable and put back all the regulators in case of an error. This
could lead to unbalanced disable or put in various cases. Furthermore
this function is also called from resume() and in this case an error
would also wrongly put back the regulator.

To fix this mess we move the initial get to the probe function and use
devm along with the bulk API. This make the code a lot simpler and
properly handle all error cases.

Change-Id: Ibd047ee7acfb86cd15fb1b3c542d7e514ee6114d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6948
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>