ARM: dts: tegra: Fix the usage of the nvidia,dsi-controller-vs property
The nvidia,dsi-controller-vs property indicates the version of the DSI
controller. So it should be set according to the SoC version, 0 for
Tegra 114 and 1 for 124. As such it makes no sense to define it at the
board level, or worse at the panel level where it is not used at all.
Cleanup this small mess by setting it in the tegra124 dtsi and
removing it everywhere else.

Finally it should be noted that this kind of thing should be handled
using different compatible strings and just dropping this useless
property, but that is for another commit.

Change-Id: I685e94beac9383d2edf810cf20f6099aae4bb50b
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6980
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>