[media] tegra: vi2: Rework the capture loop
The current capture loop doesn't allow capturing at full speed in some
scenarios. When only waiting for the buffer written sync point we
sometimes miss the next frame, leading to only half the frame rate.

The new capture loop now queues the next buffer as soon the
start-of-frame sync point triggers, and then waits for buffer to be
written before returning it. This allows the capture loop to run at
full speed in all cases.

Change-Id: I6e8d32195041b76ebca73ac333158ea7b2d99813
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7528
Reviewed-by: Julian Scheel <julian@jusst.de>