[media] tegra: vi2: Rework the capture loop
The current capture loop doesn't allow capture at full speed in some
scenario. When only waiting for the buffer written sync point we
sometimes miss the next frame, leading to only half the frame rate.

The new capture loop now queue the next buffer as soon the
start-of-frame sync point trigger, and then wait for buffer to be
written before returning it. This allow the capture loop to run at
full speed in all cases.

Change-Id: I6e8d32195041b76ebca73ac333158ea7b2d99813
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>