[media] tegra: vi2: Fix the OF nodes reference counting
When parsing the config from OF some node reference count weren't
properly decremented after use. Add the missing calls to of_node_put().

On the other hand no reference was kept for the nodes added for the
subdev async registration. So remove this of_node_put() to keep the
reference.

Change-Id: Ibb749c5a68b59ae9f6b65a18334906b961fa1af1
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7610
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>