[media] tegra: vi2: Add support for several subdev per input
Allow registering several subdevice per input, for example to support
flash devices. With this change each VI endpoint can reference up to 3
subdevices. For example:

	vi {
		port@0 {
			endpoint@0 {
				remote-endpoint = <&camera>, <&flash>;
			};
		};
	};

Of the referenced endpoint only the first one providing video
operations will be used as a sensor. The others are only registered as
subdevice and won't be used by VI device.

Change-Id: I7cb0533e02f4fbb98372ba8ccc84a089f166715e
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7612
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>