i2c: imx290: Add generic compatibility string
Add imx290 as a 3rd compatiblity string. This generic
compatibility string accepts both, b/w and color pixelcodes.

This compatibility string is necessary for boards where the type
of the sensor is flexible and thus the DTS should not restrict
whether it is a b/w or monochrome sensor.

To achieve this we rework the logic used to get/try the mbus format to
use an array lookup instead of just hardcoding a single value. This
also should make it easy to support 10 bit formats in the future if
that is ever needed.

Change-Id: Ieac89b8ef5982cf9b5ae9069dfa19fc3bd3174c0
Signed-off-by: Marc Andre <marc.andre@netline.ch>
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7613
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>