ARM: tegra: eve: Add TEC DAC&ADC and change Laser DAC VREF
Add the ltc2617 DAC for the TEC control and the ads1015 for the TEC
monitoring.

For better separation between the two ltc2617 DAC, rename the
laser DAC to ltc2617_laser.

Having a vcc supply now available for the ltc2617 DAC, change the
vref of the Laser DAC to be always on and add a new regulator for
the vcc of the laser DAC which controls the actual power supply
using GPIO.

Change-Id: I59258e359087f490059a3f6ad89e4b67504cfb74
Signed-off-by: Marc Andre <marc.andre@netline.ch>
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7619
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>