[media] tegra: vi2: Fix a crash with invalid device tree
If an enpoint reference is wrongly set and its OF node is less then 3
level deep the driver would crash. This is because we loop on
of_get_next_parent() without checking if a parent is returned.

While at it also fix the indentation of the call to
of_get_next_parent() to use tabs as it should.

Change-Id: Iece53ee7b386b238ac083aec42b090a4aab2954f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7968
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>