ARM: dts: clearview: Add support for the new adapter board
The new adapter board extend the clearview board with a video capture
PCIe device, an HDMI to analog video converter, a CAN bus and
connectors for CSI/DSI and LVDS.

To support this adapter board we only need to enable the third USB bus
for the CAN controller. As the requiered IOs are not connected without
the adapter board we can just extend the original config. Without the
adapter board it then only have one empty USB more.

Change-Id: I4b1d32b037ec66ea69346b733ed92b75e5178583
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8270
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>