regmap: Add support for data in mixed endianness
Mixed endianness is used by some i2c chips that have both 16 and 32
bits register. This happen when the 16 bit registers are access as
big-endian but the 32 bits registers are considered as 2 consecutive
16-bits register with increasing address, so little endian.

The above format is now named big-little, for 2 big endian words in
little endian oder. While at it we also add the reverse, little-big
for 2 little endian words in big endian order.

Change-Id: I81e51e81d1a7987e45f5703cfa6d55c7bd877e5f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8592
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>