misc: Add a driver for GS4910b/11b graphics clock and timing generator
The GS4910b/11b graphics clock and timing generator allow to generate
various video clock signals either genlocked to an input or in a
free-running mode.

This driver expose the most interresting registers to sysfs allowing
to read the input mode and set the output mode. If the locking
interrupts are available the attributes for the input locking state
can be polled for change.

Change-Id: Ica355c647265812d3b64f458a0ed44a30b9f3102
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8605
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>