Remove all the Werror added by nvidia
Werror is nice for developers working on a project, however as each
gcc regularily add new warnings, it just guarantee that the build will
break with some later gcc version. To avoid having to fix this kernel
for each gcc release we just remove all the Werror added by nvidia.

There is a few left in other drivers, as these drivers are generally
not used on tegra platforms we just keep them to make eventual merges
easier.

Change-Id: I2478248dfe0ef230ecc736e596c99ad0963ef44b
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8650
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>