ARM: dts: vxb: Keep the FPGA out of reset in the default config
This board has a special construct to allow programming the FPGA flash
from the main processor. A line is used to select between the normal
mode and programming mode. When in programming mode the FPGA is held
in reset and the flash can be accessed from the processor. In the
normal mode the FPGA is running and the flash is not accessible from
the processor.

This line is controlled from the flash driver to automatically switch
to the programming mode when the flash is accessed. However per
default the programming mode is enabled, this mean that the FPGA only
starts if the flash driver is loaded.

To avoid this problem we change the default to make sure the FPGA is
running even if the flash driver is not present.

Change-Id: I0e4fc47d2ed2446c5850ecb0d14f378c62c89d00
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8713
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>