usb: phy: tegra: Fix the symbols exports
The PHY driver expose an API that is used by the UDC driver,
however the build is currently broken for modules because most of
these symbols are not exported. Add the missing EXPORT_SYMBOL_GPL() on
all the non static functions.

Furthermore it turns out that there are several global functions that
are in fact only used in the generic tegra USB PHY code, so make these
static.

Change-Id: I09eb5b406f69f3578f5432d0c2976c725b7f287f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/8898
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>